backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
flask
flask-cors
pandas==2.2.2
python-dotenv==1.0.1
openai==1.51.0
requests==2.32.3
pydantic==2.8.2

flask
flask-cors
pandas
python-dotenv
openai>=1.40.0

flask==3.1.0
flask-cors==4.0.1
pandas==2.2.2
python-dotenv==1.0.1
openai==1.51.0
requests==2.32.3

geoip2
httpx==0.27.2
scikit-learn==1.5.2